move functions to loma_tools
